drm/dsi: Add set_state function to mipi_dsi_host_ops
Many DSI sink devices need the DSI host to be in a particular
state before they initialise. With the current infrastructure
the DSI host only gets powered on after all panel prepare and
bridge pre_enable calls have been made, but it also starts
video streaming at this point.

Add a new function that allows a bridge/panel to bring up
the DSI host to a defined state on request.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>